Clarify RegisterCommands method documentation
Updated the summary documentation for the `RegisterCommands` method in the `BotBuilder` class within the `TelegramBot.Builders` namespace. The new documentation specifies that this method only registers commands in the Telegram UI and that controllers need to be registered separately. It also clarifies that commands can be used without calling this method.